Validate timeout input, fix checkbox glyph, make layout responsive
- Timeout field: only single-digit keystrokes/paste reach the TextBox
  (ConnectionView.xaml.cs), and IntRangeValidationRule enforces 1-5,
  driving a red border via Validation.HasError in the shared TextBox
  style. TimeoutSeconds moved from double to int in ConnectionViewModel
  to match (whole seconds only), with MainViewModel updated accordingly.
- CheckBox glyph was neither centered nor sized right: the Path had no
  explicit Width/Height/Stretch/alignment, so it rendered at its raw
  Data-coordinate offset instead of centered in the 18x18 box. Fixed
  with Stretch="Uniform" plus explicit size and Center alignment, and
  trimmed the stroke down slightly.
- ConnectionView and FirmwareView's options row now use WrapPanel
  instead of fixed-width Grid columns / a plain StackPanel, so field
  groups drop to their own line rather than clipping or overlapping
  when the window is narrowed. FlashLogView and MainWindow's outer grid
  were already responsive (star-sized rows/columns).

+++ src/NetBootloader.App/ViewModels/ConnectionViewModel.cs
@@ -16,7 +16,7 @@ public sealed partial class ConnectionViewModel : ObservableObject
         AvailablePorts = new ObservableCollection<string>();
         BaudRates = new ObservableCollection<int>(CommonBaudRates);
         SelectedBaudRate = 460800;
-        TimeoutSeconds = 1.0;
+        TimeoutSeconds = 1;
         RefreshPorts();
     }
 
@@ -30,8 +30,9 @@ public sealed partial class ConnectionViewModel : ObservableObject
     [ObservableProperty]
     private int _selectedBaudRate;
 
+    /// <summary>Serial read timeout, in whole seconds. Valid range is 1-5, enforced by the view's binding.</summary>
     [ObservableProperty]
-    private double _timeoutSeconds;
+    private int _timeoutSeconds;
 
     [RelayCommand]
     private void RefreshPorts()
